Major update. PST works, PST printer works. AST printer works. AST needs work.

+++ Sym.cpp
@@ -46,7 +46,7 @@ void Sym::addOcc(int _ln, int _ix) {
 }
 
 int Sym::getNumOccs() {
-    return occs.size();
+    return static_cast<int>(occs.size());
 }
 
 int Sym::getOccLineNum(int occ) {
